+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+# CHANGELOG.md
 # Change Log
 
 All notable changes to the "select-php-method" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.0.2] - 2024-04-20
 
-## [Unreleased]
+### Fixed
+- Selección incorrecta de métodos con llaves anidadas o comentarios internos
+- Mejorado el reconocimiento de métodos con PHPDoc
+- Corrección del cierre de llaves del método
 
-- Initial release
+## [0.0.1] - 2024-04-15
+
+### Added
+- Función inicial para seleccionar y copiar métodos PHP